mpv audio fix script
This adds several different scripts. They all disable audio for a short time (0.4s - changeable) and reenable it. This is needed on some AVRs so they do not get overflown with audio data. With a short break, the audio gets streamed freshly again.

There are three types:
reinitonseek => sets the audio track to 0 and back again
reinitaoonseek => checks current ao and sets it to null and back again (errors out when used too aggressively)
reinitaoonseekV2 => desired ao is hardcoded (wasapi), sets to null and back again (not erroring out)